Modify: Upgrade lots of dependencies...

+++ ZMain/ImkGithubApp/src/main/java/net/imknown/imkgithub/mvp/BaseMvpPresenterActivity.kt
@@ -4,7 +4,7 @@ import org.jetbrains.anko.toast
 
 abstract class BaseMvpPresenterActivity : BaseActivity(), BaseMvpView {
 
-    override fun showMessage(@MessageTypeDef messageType: Long, message: String) {
+    override fun showMessage(@MessageTypeDef messageType: Int, message: String) {
         if (messageType == BaseMvpView.MessageType.Toast) {
             toast(message)
         }
